defaul-graphical-interface

+++ src/Controller/ClasseController.php
@@ -20,6 +20,7 @@ use Symfony\Component\Security\Core\Security;
 class ClasseController extends AbstractController
 {
     /**
+     * @Route("/")
      * @Route("/classes", name="classes")
      *
      * @param ClasseRepository $classeRepository
@@ -84,20 +85,26 @@ class ClasseController extends AbstractController
     }
 
     /**
-     * @Route("/classe/edit/{id}", name="classe_edit")
+     * @Route("/classe/edit/{id}/{redirect}", name="classe_edit", defaults={"redirect"=null})
      * @IsGranted("ROLE_CLASS_EDIT", statusCode=404, message="Not found")
      *
      * @param Request $request
      * @param Classe $classe
      * @return Response
      */
-    public function edit(Request $request, Classe $classe): Response
+    public function edit(Request $request, Classe $classe, ?String $redirect): Response
     {
         $form = $this->createForm(ClasseType::class, $classe);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
+
+            if(!is_null($redirect)) {
+                $redirect = json_decode(base64_decode($redirect), true);
+                return $this->redirectToRoute($redirect['route'], $redirect["params"]);
+            }
+
             return $this->redirectToRoute('classes');
         }
 
@@ -123,11 +130,15 @@ class ClasseController extends AbstractController
             return $this->json(['message' => 'Invalid csrf token'], 201);
         }
 
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($classe);
-        $entityManager->flush();
+        // Checking if the class has activities in it.
+        if(!($classe->getActivites()) > 0) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($classe);
+            $entityManager->flush();
+            return $this->json(['message' => 'Classe deleted'], 200);
+        }
 
-        return $this->json(['message' => 'Classe deleted'], 200);
+        return $this->json(['error' => true], 200);
     }
 
 
@@ -240,7 +251,11 @@ class ClasseController extends AbstractController
     public function viewClassUsers(Classe $classe)
     {
         return $this->render('users/index.html.twig', [
-           'users' => $classe->getUsers(),
+           'users'   => $classe->getUsers(),
+           'redirect' => base64_encode(json_encode([
+               'route'  => 'classe_view_users',
+               'params' => ['id' => $classe->getId()],
+           ])),
         ]);
     }
 
@@ -255,6 +270,10 @@ class ClasseController extends AbstractController
     {
         return $this->render('eleve/index.html.twig', [
             'eleves' => $classe->getEleves(),
+            'redirect' => base64_encode(json_encode([
+                'route'  => 'classe_view_students',
+                'params' => ['id' => $classe->getId()],
+            ])),
         ]);
     }
